+++ src/components/DefaultNavbar.js
@@ -11,6 +11,7 @@ import NavLink from '@material-tailwind/react/NavLink';
 import Dropdown from '@material-tailwind/react/Dropdown';
 import DropdownItem from '@material-tailwind/react/DropdownItem';
 import Icon from '@material-tailwind/react/Icon';
+import Button from '@material-tailwind/react/Button';
 
 export default function DefaultNavbar() {
     const [openNavbar, setOpenNavbar] = useState(false);
@@ -115,6 +116,19 @@ export default function DefaultNavbar() {
                             >
                                 Issues
                             </NavLink>
+                            <a
+                                href="https://www.creative-tim.com/product/material-tailwind-kit-react"
+                                target="_blank"
+                                rel="noreferrer"
+                            >
+                                <Button
+                                    color="transparent"
+                                    className="bg-white text-black ml-4"
+                                    ripple="dark"
+                                >
+                                    Free Download
+                                </Button>
+                            </a>
                         </div>
                     </Nav>
                 </NavbarCollapse>